Implement refined inlining
This will try to determine whether a bug is definitely present or definitely not present in the input.
If all monitors report a bug, then it's reported and the searach continues. If all monitors report
no bugs present, then exploration continues.
If *some* monitors report a bug for a region, inlining is used to determine whether it's present or not.