cfgPrinter: Remove controling the visited states
We have no idea whether eba is on a new path, even if we see the same
step again.  It is better to let EBA handle the fixpoint.  I have no
idea how it happens, but this is a blind trust.  This way we are not
loosing monitors when revisiting the same node again.